[py3] Use list comprehension when building opts
In python, set() is not ordered. When testing the project, the unit
tests would allways succeed for python 2.x but most often fail for python
3.x (4 fails out of 5 run).

It seems that although set() is not ordered in python 2.x, it always
create the same output list (which has been harcoded in the unit test).

However, with python3 the set() will never have the same order, which
would make the unit test fails most of the time, except when being
lucky.

To prevent shuffling the interpreter options around, this commit removes
the use of set() and use list() and list comprehension to behave like an
(ordered) set()

(took me ages to find the bug...)